move enemy class to killer subdir
