+++ src/objects/projectile/Bullet.ts
@@ -1,5 +1,5 @@
 import { Vector } from "../../utils/Vector";
-import { Enemy } from "../Enemy";
+import { Enemy } from "../killers/Enemy";
 import { Projectile } from "./Projectile";
 import * as PIXI from "pixi.js";
 import sound from "pixi-sound";
