+++ src/objects/projectile/LightningBolt.ts
@@ -1,5 +1,5 @@
 import Position from "src/ts/types/Position";
-import { Enemy } from "../Enemy";
+import { Enemy } from "../killers/Enemy";
 import { Projectile } from "./Projectile";
 import * as PIXI from "pixi.js";
 import sound from "pixi-sound";
